Refactor models, factor out freebase
